Move TokenStateTranslator back to state.migration package
